// src/types/skills.ts
export type Skill = {
    id: string
    code: string
    name: string
    imageUrl: string
    status: number
    createdDate: string
    lastUpdate: string
    robotModelId: string
}

export type SkillModal = {
    code: string
    name: string
    imageUrl: string
    status: number
    robotModelId: string
}